+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## NEXT
+
+- Added `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.cache_nodes` and
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.cache_nodes` to avoid having to loop
+  through the nodes of the graph multiple times.
+
 ## 0.4.6
 
 - Added the Boolean keyword argument `include_all_parents` to